Move aliasutils into AddAliasInfo
